Revert to old way of setting github repo and org
While it turns out that the SourceLink build bits do all the work of deducing the repo URL for you, they turn out not to do any of that work during pack-only builds. (So if --no-build is set during a pack, that turns off all that behaviour.)

So although it looked like we could piggy-back successfully on all of that work it does, it turns out not to work in practice.

It's possible we could find ways to use the components that sourcelink uses to do all this, but it's quicker to revert to horrible regex at this point.